refactor(plugin): move standalone DTOs and domain models into packages
Part of #730 (epic #728).

- api/dto: FlipFinderResponse, ActiveFlipsResponse, CompletedFlipsResponse,
  FlipStatisticsResponse, BankSnapshotResponse, BankSnapshotStatusResponse,
  BlocklistsResponse, BlocklistSummary, FlipAdjustmentResponse,
  TimeframeFlipFinderResponse, OfferAdviceRequest, OfferAdviceResponse,
  OfferAdviceBatchResponse, OfferAdviceResult, DumpEvent.
- domain/flip: FlipRecommendation, FlipAnalysis, ActiveFlip, CompletedFlip.
- domain/offer: TrackedOffer, PendingOrder, OfferDisposition, OfferAction.

Pure file moves with package decls + imports updated across consumers.
No behavior change.
