+++ src/main/java/com/flipsmart/DumpAlertService.java
@@ -1,4 +1,5 @@
 package com.flipsmart;
+import com.flipsmart.api.dto.DumpEvent;
 
 import lombok.extern.slf4j.Slf4j;
 import net.runelite.api.ChatMessageType;
