+++ src/main/java/com/flipsmart/api/dto/CompletedFlipsResponse.java
@@ -1,4 +1,5 @@
-package com.flipsmart;
+package com.flipsmart.api.dto;
+import com.flipsmart.domain.flip.CompletedFlip;
 
 import com.google.gson.annotations.SerializedName;
 import lombok.Data;
