+++ src/main/java/com/flipsmart/api/dto/OfferAdviceBatchResponse.java
@@ -1,4 +1,4 @@
-package com.flipsmart;
+package com.flipsmart.api.dto;
 
 import java.util.List;
 import lombok.Data;
