+++ src/main/java/com/flipsmart/api/dto/OfferAdviceResponse.java
@@ -1,4 +1,5 @@
-package com.flipsmart;
+package com.flipsmart.api.dto;
+import com.flipsmart.domain.offer.OfferAction;
 
 import com.google.gson.annotations.SerializedName;
 import lombok.Data;
