+++ src/main/java/com/flipsmart/domain/offer/OfferDisposition.java
@@ -1,4 +1,4 @@
-package com.flipsmart;
+package com.flipsmart.domain.offer;
 
 public enum OfferDisposition
 {
