+++ src/test/java/com/flipsmart/OfferActionBodyTest.java
@@ -1,4 +1,5 @@
 package com.flipsmart;
+import com.flipsmart.api.dto.OfferAdviceRequest;
 
 import com.google.gson.JsonObject;
 import org.junit.Test;
