+++ src/test/java/com/flipsmart/OfferAdviceResponseTest.java
@@ -1,4 +1,6 @@
 package com.flipsmart;
+import com.flipsmart.domain.offer.OfferAction;
+import com.flipsmart.api.dto.OfferAdviceResponse;
 
 import com.google.gson.Gson;
 import org.junit.Test;
