+++ src/test/java/com/flipsmart/PlayerSessionTest.java
@@ -1,4 +1,5 @@
 package com.flipsmart;
+import com.flipsmart.domain.offer.TrackedOffer;
 
 import org.junit.Test;
